requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
Authlib==1.3.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-rest-knox==4.2.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.36.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
